+++ src/main/java/com/fouristhenumber/utilitiesinexcess/UtilitiesInExcess.java
@@ -6,11 +6,11 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.util.WeightedRandomChestContent;
 import net.minecraftforge.common.ChestGenHooks;
 import net.minecraftforge.common.ForgeChunkManager;
-import net.minecraftforge.common.MinecraftForge;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import com.fouristhenumber.utilitiesinexcess.common.items.ItemInvertedIngot;
 import com.fouristhenumber.utilitiesinexcess.common.recipe.RecipeLoader;
 import com.fouristhenumber.utilitiesinexcess.common.renderers.BlackoutCurtainsRenderer;
 import com.fouristhenumber.utilitiesinexcess.common.renderers.LapisAetheriusRenderer;
@@ -56,8 +56,7 @@ import com.fouristhenumber.utilitiesinexcess.compat.Mods;
 import com.fouristhenumber.utilitiesinexcess.compat.crafttweaker.EnderLocusCraftTweakerSupport;
 import com.fouristhenumber.utilitiesinexcess.compat.tinkers.TinkersCompat;
 import com.fouristhenumber.utilitiesinexcess.config.OtherConfig;
-import com.fouristhenumber.utilitiesinexcess.utils.FMLEventHandler;
-import com.fouristhenumber.utilitiesinexcess.utils.ForgeEventHandler;
+import com.fouristhenumber.utilitiesinexcess.config.items.InversionConfig;
 import com.fouristhenumber.utilitiesinexcess.utils.PinkFuelHelper;
 import com.fouristhenumber.utilitiesinexcess.utils.PumpChunkLoadingCallback;
 
@@ -108,10 +107,11 @@ public class UtilitiesInExcess {
         proxy.init(event);
         RecipeLoader.run();
 
-        MinecraftForge.EVENT_BUS.register(new ForgeEventHandler());
-        FMLCommonHandler.instance()
-            .bus()
-            .register(new FMLEventHandler());
+        if (InversionConfig.enableInvertedIngot) {
+            FMLCommonHandler.instance()
+                .bus()
+                .register(new ItemInvertedIngot.FMLEvents());
+        }
 
         GameRegistry.registerTileEntity(TileEntitySpike.class, "TileEntitySpikeUIE");
         GameRegistry.registerTileEntity(TileEntityRedstoneClock.class, "TileEntityRedstoneClockUIE");
