+++ src/main/java/com/fouristhenumber/utilitiesinexcess/common/blocks/BlockGigaTorch.java
@@ -7,6 +7,8 @@ import java.util.List;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockContainer;
 import net.minecraft.block.material.Material;
+import net.minecraft.entity.EnumCreatureType;
+import net.minecraft.entity.monster.EntitySlime;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemBlock;
 import net.minecraft.item.ItemStack;
@@ -14,10 +16,16 @@ import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.AxisAlignedBB;
 import net.minecraft.util.StatCollector;
 import net.minecraft.world.World;
+import net.minecraftforge.event.entity.living.LivingSpawnEvent;
 
+import com.fouristhenumber.utilitiesinexcess.UtilitiesInExcess;
 import com.fouristhenumber.utilitiesinexcess.common.tileentities.TileEntityGigaTorch;
 import com.fouristhenumber.utilitiesinexcess.config.blocks.BlockConfig;
 import com.gtnewhorizon.gtnhlib.client.model.ModelISBRH;
+import com.gtnewhorizon.gtnhlib.eventbus.EventBusSubscriber;
+
+import cpw.mods.fml.common.eventhandler.Event;
+import cpw.mods.fml.common.eventhandler.SubscribeEvent;
 
 public class BlockGigaTorch extends BlockContainer {
 
@@ -83,6 +91,44 @@ public class BlockGigaTorch extends BlockContainer {
         public void addInformation(ItemStack stack, EntityPlayer player, List<String> tooltip, boolean bool) {
             tooltip.add(
                 StatCollector.translateToLocalFormatted("tile.giga_torch.desc", BlockConfig.gigaTorch.gigaTorchRange));
+        }
+    }
+
+    @SuppressWarnings("unused")
+    @EventBusSubscriber
+    public static class Events {
+
+        @EventBusSubscriber.Condition
+        public static boolean shouldSubscribe() {
+            return BlockConfig.gigaTorch.enableGigaTorch || BlockConfig.chandelier.enableChandelier;
+        }
+
+        // context https://github.com/GTNewHorizons/GT5-Unofficial/pull/905
+        @SubscribeEvent
+        public static void denySpawn(LivingSpawnEvent.CheckSpawn event) {
+            if (event.getResult() == Event.Result.DENY) return;
+
+            if (event.entityLiving instanceof EntitySlime slime && !slime.hasCustomNameTag()
+                && event.getResult() == Event.Result.ALLOW) {
+                event.setResult(Event.Result.DEFAULT);
+            }
+
+            if (event.getResult() == Event.Result.ALLOW) {
+                return;
+            }
+
+            if (event.entityLiving.isCreatureType(EnumCreatureType.monster, false)) {
+                if (UtilitiesInExcess.proxy.mobSpawnBlockChecks.isInRange(
+                    event.entity.worldObj.provider.dimensionId,
+                    event.entity.posX,
+                    event.entity.posY,
+                    event.entity.posZ)) {
+                    if (event.entityLiving instanceof EntitySlime slime) {
+                        slime.setCustomNameTag("DoNotSpawnSlimes");
+                    }
+                    event.setResult(Event.Result.DENY);
+                }
+            }
         }
     }
 }
