+++ src/main/java/com/fouristhenumber/utilitiesinexcess/common/items/ItemInversionSigilActive.java
@@ -29,7 +29,6 @@ import net.minecraft.util.StatCollector;
 import net.minecraft.world.EnumDifficulty;
 import net.minecraft.world.World;
 import net.minecraftforge.common.DimensionManager;
-import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.common.util.ForgeDirection;
 import net.minecraftforge.event.entity.living.LivingDeathEvent;
 import net.minecraftforge.event.entity.living.LivingSpawnEvent;
@@ -41,8 +40,8 @@ import com.fouristhenumber.utilitiesinexcess.ModItems;
 import com.fouristhenumber.utilitiesinexcess.common.entities.EntitySiegeProperty;
 import com.fouristhenumber.utilitiesinexcess.config.items.InversionConfig;
 import com.fouristhenumber.utilitiesinexcess.utils.ItemStackBaseCompare;
+import com.gtnewhorizon.gtnhlib.eventbus.EventBusSubscriber;
 
-import cpw.mods.fml.common.FMLCommonHandler;
 import cpw.mods.fml.common.FMLLog;
 import cpw.mods.fml.common.eventhandler.Event;
 import cpw.mods.fml.common.eventhandler.EventPriority;
@@ -55,15 +54,15 @@ public class ItemInversionSigilActive extends Item {
 
     public static final String DURABILITY_NBT_KEY = "RemainingUses";
     private static final int BEACON_SEARCH_RADIUS = 6;
-    private final int[][] LIGHTNING_POSITIONS = { { 0, 0 }, { -5, 0 }, { 5, 0 }, { 0, -5 }, { 0, 5 } };
+    private static final int[][] LIGHTNING_POSITIONS = { { 0, 0 }, { -5, 0 }, { 5, 0 }, { 0, -5 }, { 0, 5 } };
 
-    private HashSet<ItemStackBaseCompare> CHEST_NORTH_CONTENTS = new HashSet<>();
+    private static HashSet<ItemStackBaseCompare> CHEST_NORTH_CONTENTS = new HashSet<>();
 
-    private HashSet<ItemStackBaseCompare> CHEST_EAST_CONTENTS = new HashSet<>();
+    private static HashSet<ItemStackBaseCompare> CHEST_EAST_CONTENTS = new HashSet<>();
 
-    private HashSet<ItemStackBaseCompare> CHEST_SOUTH_CONTENTS = new HashSet<>();
+    private static HashSet<ItemStackBaseCompare> CHEST_SOUTH_CONTENTS = new HashSet<>();
 
-    private HashSet<ItemStackBaseCompare> CHEST_WEST_CONTENTS = new HashSet<>();
+    private static HashSet<ItemStackBaseCompare> CHEST_WEST_CONTENTS = new HashSet<>();
 
     public ItemInversionSigilActive() {
         super();
@@ -91,12 +90,6 @@ public class ItemInversionSigilActive extends Item {
             ForgeDirection.WEST,
             InversionConfig.westChestValidItems,
             InversionConfig.westChestRequiredItems);
-
-        ItemInversionSigilActiveEvents eventHandler = new ItemInversionSigilActiveEvents();
-        MinecraftForge.EVENT_BUS.register(eventHandler);
-        FMLCommonHandler.instance()
-            .bus()
-            .register(eventHandler);
     }
 
     private int parseItemMetaFromString(String string) {
@@ -166,11 +159,11 @@ public class ItemInversionSigilActive extends Item {
         return validChestContents;
     }
 
-    private EntitySiegeProperty getProperties(EntityPlayer player) {
+    private static EntitySiegeProperty getProperties(EntityPlayer player) {
         return (EntitySiegeProperty) player.getExtendedProperties(EntitySiegeProperty.PROP_KEY);
     }
 
-    private List<EntityPlayer> getSiegePlayers() {
+    private static List<EntityPlayer> getSiegePlayers() {
         List<EntityPlayer> siegePlayers = new ArrayList<>();
         if (DimensionManager.getWorld(1) == null) {
             return siegePlayers;
@@ -184,7 +177,7 @@ public class ItemInversionSigilActive extends Item {
         return siegePlayers;
     }
 
-    private boolean checkSpiral(World world, int x, int y, int z) {
+    private static boolean checkSpiral(World world, int x, int y, int z) {
         ForgeDirection[] directions = { ForgeDirection.NORTH, ForgeDirection.EAST, ForgeDirection.SOUTH,
             ForgeDirection.WEST };
         boolean clockwise = true;
@@ -223,7 +216,7 @@ public class ItemInversionSigilActive extends Item {
         return true;
     }
 
-    private boolean checkSpiralPart(World world, int x, int y, int z, ForgeDirection direction, Block block) {
+    private static boolean checkSpiralPart(World world, int x, int y, int z, ForgeDirection direction, Block block) {
         int[] SPIRAL_SEGMENT_LENGTHS = { 3, 5, 7, 8 };
 
         for (int segmentLength : SPIRAL_SEGMENT_LENGTHS) {
@@ -244,7 +237,7 @@ public class ItemInversionSigilActive extends Item {
         return true;
     }
 
-    private ForgeDirection rotateClockwise(ForgeDirection direction) {
+    private static ForgeDirection rotateClockwise(ForgeDirection direction) {
         return switch (direction) {
             case NORTH -> ForgeDirection.EAST;
             case EAST -> ForgeDirection.SOUTH;
@@ -255,7 +248,7 @@ public class ItemInversionSigilActive extends Item {
         };
     }
 
-    private ForgeDirection rotateCounterClockwise(ForgeDirection direction) {
+    private static ForgeDirection rotateCounterClockwise(ForgeDirection direction) {
         return switch (direction) {
             case NORTH -> ForgeDirection.WEST;
             case EAST -> ForgeDirection.NORTH;
@@ -266,7 +259,7 @@ public class ItemInversionSigilActive extends Item {
         };
     }
 
-    private void startSiege(World world, int beaconX, int beaconY, int beaconZ, EntityPlayer player) {
+    private static void startSiege(World world, int beaconX, int beaconY, int beaconZ, EntityPlayer player) {
         EntitySiegeProperty source = getProperties(player);
         source.beaconSpawnX = beaconX;
         source.beaconSpawnY = beaconY;
@@ -290,7 +283,7 @@ public class ItemInversionSigilActive extends Item {
         }
     }
 
-    private void endSiege(boolean won, EntityPlayer player) {
+    private static void endSiege(boolean won, EntityPlayer player) {
         EntitySiegeProperty source = getProperties(player);
         source.siege = false;
         source.siegeMobsKilled = 0;
@@ -303,7 +296,7 @@ public class ItemInversionSigilActive extends Item {
         if (won) {
             for (int i = 0; i < player.inventory.getSizeInventory(); i++) {
                 ItemStack stack = player.inventory.getStackInSlot(i);
-                if (stack != null && stack.getItem() == this) {
+                if (stack != null && stack.getItem() == ModItems.INVERSION_SIGIL_ACTIVE.get()) {
                     player.inventory.setInventorySlotContents(i, ModItems.PSEUDO_INVERSION_SIGIL.newItemStack());
                     break;
                 }
@@ -311,7 +304,8 @@ public class ItemInversionSigilActive extends Item {
         }
     }
 
-    private boolean checkChest(TileEntityChest chest, HashSet<ItemStackBaseCompare> itemsToCheck, int requiredAmount) {
+    private static boolean checkChest(TileEntityChest chest, HashSet<ItemStackBaseCompare> itemsToCheck,
+        int requiredAmount) {
         HashSet<ItemStackBaseCompare> confirmedItems = new HashSet<>();
         for (int i = 0; i < chest.getSizeInventory(); i++) {
             ItemStack stack = chest.getStackInSlot(i);
@@ -324,7 +318,7 @@ public class ItemInversionSigilActive extends Item {
         return confirmedItems.size() >= requiredAmount;
     }
 
-    private boolean checkChestInDirection(ForgeDirection direction, int beaconX, int beaconY, int beaconZ,
+    private static boolean checkChestInDirection(ForgeDirection direction, int beaconX, int beaconY, int beaconZ,
         World world) {
         HashSet<ItemStackBaseCompare> contents;
         int requiredAmount;
@@ -425,7 +419,7 @@ public class ItemInversionSigilActive extends Item {
         return true;
     }
 
-    private boolean isValidRitualBeacon(World world, int beaconX, int beaconY, int beaconZ) {
+    private static boolean isValidRitualBeacon(World world, int beaconX, int beaconY, int beaconZ) {
         boolean dimensionOk = (world.provider.dimensionId == 1);
 
         return dimensionOk && world.getBlock(beaconX, beaconY, beaconZ) == Blocks.beacon
@@ -436,7 +430,8 @@ public class ItemInversionSigilActive extends Item {
             && checkChestInDirection(ForgeDirection.WEST, beaconX, beaconY, beaconZ, world);
     }
 
-    public class ItemInversionSigilActiveEvents {
+    @SuppressWarnings("unused")
+    public static class FMLEvents {
 
         @SubscribeEvent(priority = EventPriority.NORMAL)
         public void whenServerTick(TickEvent.ServerTickEvent event) {
@@ -517,9 +512,19 @@ public class ItemInversionSigilActive extends Item {
                 endSiege(false, event.player);
             }
         }
+    }
+
+    @SuppressWarnings("unused")
+    @EventBusSubscriber
+    public static class Events {
+
+        @EventBusSubscriber.Condition
+        public static boolean shouldSubscribe() {
+            return InversionConfig.enableInversionSigil;
+        }
 
         @SubscribeEvent(priority = EventPriority.NORMAL)
-        public void whenEndermanSpawn(LivingSpawnEvent.CheckSpawn event) {
+        public static void whenEndermanSpawn(LivingSpawnEvent.CheckSpawn event) {
             if (event.world.provider.dimensionId == 1 && event.entity instanceof EntityEnderman
                 && !getSiegePlayers().isEmpty()) {
                 event.setResult(Event.Result.DENY);
@@ -527,7 +532,7 @@ public class ItemInversionSigilActive extends Item {
         }
 
         @SubscribeEvent(priority = EventPriority.NORMAL)
-        public void onLivingDeath(LivingDeathEvent event) {
+        public static void onLivingDeath(LivingDeathEvent event) {
 
             World world = event.entityLiving.worldObj;
 
@@ -573,7 +578,7 @@ public class ItemInversionSigilActive extends Item {
 
             if (source.siege) return; // Cannot start a second siege while in a siege.
             if (world.difficultySetting == EnumDifficulty.PEACEFUL) return;
-            if (!player.inventory.hasItem(ItemInversionSigilActive.this)) return;
+            if (!player.inventory.hasItem(ModItems.INVERSION_SIGIL_ACTIVE.get())) return;
 
             int radius = BEACON_SEARCH_RADIUS;
             int mobX = (int) Math.floor(event.entityLiving.posX);
