+++ src/main/java/com/fouristhenumber/utilitiesinexcess/common/items/tools/ItemDestructionPickaxe.java
@@ -12,11 +12,14 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.util.EnumChatFormatting;
 import net.minecraft.util.StatCollector;
 import net.minecraft.world.World;
+import net.minecraftforge.event.world.BlockEvent;
 
 import com.fouristhenumber.utilitiesinexcess.config.items.unstabletools.DestructionPickaxeConfig;
 import com.gtnewhorizon.gtnhlib.api.ITranslucentItem;
+import com.gtnewhorizon.gtnhlib.eventbus.EventBusSubscriber;
 
 import akka.japi.Pair;
+import cpw.mods.fml.common.eventhandler.SubscribeEvent;
 
 public class ItemDestructionPickaxe extends ItemPickaxe implements ITranslucentItem {
 
@@ -95,4 +98,26 @@ public class ItemDestructionPickaxe extends ItemPickaxe implements ITranslucentI
         return super.showDurabilityBar(stack);
     }
     //
+
+    @SuppressWarnings("unused")
+    @EventBusSubscriber
+    public static class Events {
+
+        @EventBusSubscriber.Condition
+        public static boolean shouldSubscribe() {
+            return DestructionPickaxeConfig.enable;
+        }
+
+        @SubscribeEvent
+        public static void onBlockBroken(BlockEvent.HarvestDropsEvent event) {
+            if (!DestructionPickaxeConfig.voidMinedBlock) return;
+            if (event.harvester == null) return;
+            if (event.harvester.getHeldItem() == null) return;
+
+            if (event.harvester.getHeldItem()
+                .getItem() instanceof ItemDestructionPickaxe) {
+                event.drops.clear();
+            }
+        }
+    }
 }
