+++ src/main/java/com/fouristhenumber/utilitiesinexcess/common/renderers/GloveRenderer.java
@@ -2,6 +2,7 @@ package com.fouristhenumber.utilitiesinexcess.common.renderers;
 
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.ScaledResolution;
+import net.minecraft.client.renderer.RenderHelper;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.client.renderer.texture.TextureMap;
 import net.minecraft.entity.player.EntityPlayer;
@@ -11,12 +12,15 @@ import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.client.IItemRenderer;
 
 import org.lwjgl.opengl.GL11;
+import org.lwjgl.opengl.GL12;
 
 import com.fouristhenumber.utilitiesinexcess.ModItems;
 import com.fouristhenumber.utilitiesinexcess.common.items.ItemGlove;
 import com.fouristhenumber.utilitiesinexcess.utils.RenderableCube;
 import com.gtnewhorizon.gtnhlib.util.ItemRenderUtil;
 
+import cpw.mods.fml.common.eventhandler.SubscribeEvent;
+import cpw.mods.fml.common.gameevent.TickEvent;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
 
@@ -167,5 +171,46 @@ public class GloveRenderer implements IItemRenderer {
         t.draw();
 
         GL11.glPopMatrix();
+    }
+
+    @SuppressWarnings("unused")
+    public static class FMLEvents {
+
+        @SubscribeEvent
+        public void tickRender(TickEvent.RenderTickEvent event) {
+            if (!ItemGlove.isUsingGlove(Minecraft.getMinecraft().thePlayer)) return;
+
+            // Setup hud rendering
+            // Move if new hud stuff is ever added
+            if (event.phase == TickEvent.Phase.END) {
+                Minecraft mc = Minecraft.getMinecraft();
+                if (!(mc.currentScreen == null && mc.theWorld != null
+                    && Minecraft.isGuiEnabled()
+                    && !mc.gameSettings.keyBindPlayerList.getIsKeyPressed())) return;
+
+                GL11.glPushMatrix();
+
+                boolean hasBlending = GL11.glGetBoolean(GL11.GL_BLEND);
+                boolean hasDepthTest = GL11.glGetBoolean(GL11.GL_DEPTH_TEST);
+                int boundTexIndex = GL11.glGetInteger(GL11.GL_TEXTURE_BINDING_2D);
+                GL11.glPushAttrib(GL11.GL_CURRENT_BIT);
+
+                GL11.glDisable(GL12.GL_RESCALE_NORMAL);
+                RenderHelper.disableStandardItemLighting();
+                GL11.glDisable(GL11.GL_LIGHTING);
+                GL11.glDisable(GL11.GL_DEPTH_TEST);
+
+                GloveRenderer.renderGloveHudIcon();
+
+                if (hasBlending) GL11.glEnable(GL11.GL_BLEND);
+                else GL11.glDisable(GL11.GL_BLEND);
+                if (hasDepthTest) GL11.glEnable(GL11.GL_DEPTH_TEST);
+                else GL11.glDisable(GL11.GL_DEPTH_TEST);
+                GL11.glBindTexture(GL11.GL_TEXTURE_2D, boundTexIndex);
+
+                GL11.glPopAttrib();
+                GL11.glPopMatrix();
+            }
+        }
     }
 }
