+++ src/main/java/com/fouristhenumber/utilitiesinexcess/compat/findit/FindItHelper.java
@@ -7,11 +7,16 @@ import net.minecraft.client.entity.EntityClientPlayerMP;
 import net.minecraft.world.ChunkPosition;
 import net.minecraftforge.client.event.RenderWorldLastEvent;
 
+import com.fouristhenumber.utilitiesinexcess.compat.Mods;
+import com.gtnewhorizon.gtnhlib.eventbus.EventBusSubscriber;
 import com.gtnh.findit.fx.EntityHighlighter;
 import com.gtnh.findit.util.ClientFinderHelperUtils;
 
 import cpw.mods.fml.common.eventhandler.SubscribeEvent;
+import cpw.mods.fml.relauncher.Side;
 
+@SuppressWarnings("unused")
+@EventBusSubscriber(side = Side.CLIENT)
 public class FindItHelper {
 
     public static EntityHighlighter entityHighlighter;
@@ -28,8 +33,13 @@ public class FindItHelper {
         ClientFinderHelperUtils.rotateViewHelper(player, targets);
     }
 
+    @EventBusSubscriber.Condition
+    public static boolean shouldSubscribe() {
+        return Mods.FindIt.isLoaded();
+    }
+
     @SubscribeEvent
-    public void onRenderWorldLast(RenderWorldLastEvent event) {
+    public static void onRenderWorldLast(RenderWorldLastEvent event) {
         FindItHelper.entityHighlighter.renderHighlightedEntities(event);
     }
 }
