+++ src/main/java/com/fouristhenumber/utilitiesinexcess/compat/nei/NEIConfig.java
@@ -5,6 +5,8 @@ import net.minecraft.util.StatCollector;
 import com.fouristhenumber.utilitiesinexcess.ModBlocks;
 import com.fouristhenumber.utilitiesinexcess.Tags;
 import com.fouristhenumber.utilitiesinexcess.UtilitiesInExcess;
+import com.fouristhenumber.utilitiesinexcess.compat.Mods;
+import com.gtnewhorizon.gtnhlib.eventbus.EventBusSubscriber;
 
 import codechicken.nei.api.API;
 import codechicken.nei.api.IConfigureNEI;
@@ -12,6 +14,7 @@ import codechicken.nei.event.NEIRegisterHandlerInfosEvent;
 import codechicken.nei.recipe.HandlerInfo;
 import cpw.mods.fml.common.event.FMLInterModComms;
 import cpw.mods.fml.common.eventhandler.SubscribeEvent;
+import cpw.mods.fml.relauncher.Side;
 
 @SuppressWarnings("unused")
 public class NEIConfig implements IConfigureNEI {
@@ -48,13 +51,23 @@ public class NEIConfig implements IConfigureNEI {
                 + "@ender_locus_recipes");
     }
 
-    @SubscribeEvent
-    public void registerHandlerInfo(NEIRegisterHandlerInfosEvent event) {
-        event.registerHandlerInfo(
-            new HandlerInfo.Builder("ender_locus_recipes", UtilitiesInExcess.MODNAME, UtilitiesInExcess.MODID)
-                .setHeight(140)
-                .setWidth(166)
-                .setDisplayStack(ModBlocks.ENDER_LOCUS.newItemStack())
-                .build());
+    @SuppressWarnings("unused")
+    @EventBusSubscriber(side = Side.CLIENT)
+    public static class Events {
+
+        @EventBusSubscriber.Condition
+        public static boolean shouldSubscribe() {
+            return Mods.NEI.isLoaded();
+        }
+
+        @SubscribeEvent
+        public static void registerHandlerInfo(NEIRegisterHandlerInfosEvent event) {
+            event.registerHandlerInfo(
+                new HandlerInfo.Builder("ender_locus_recipes", UtilitiesInExcess.MODNAME, UtilitiesInExcess.MODID)
+                    .setHeight(140)
+                    .setWidth(166)
+                    .setDisplayStack(ModBlocks.ENDER_LOCUS.newItemStack())
+                    .build());
+        }
     }
 }
